% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceMeasure.R
\name{performanceMeasure}
\alias{performanceMeasure}
\title{Performance Measure function}
\usage{
performanceMeasure(actual, predicted)
}
\arguments{
\item{actual}{and predicted values}
}
\value{
object with various performance measures
}
\description{
This function allows you to measure performance of actual and predicted values from a model
}
\examples{
performanceMeasure()
}
\keyword{confusion}
\keyword{matrix}

